+++ tests/syntax/gt_or_eq_3092/syntax_test_issue_62_fix_comparison_operator.php
@@ -1,0 +1,29 @@
+// SYNTAX TEST "Packages/php-grammar/PHP.tmLanguage"
+<?php
+
+$comparison === 1;
+//          ^ keyword.operator.comparison
+//           ^ keyword.operator.comparison
+//            ^ keyword.operator.comparison
+
+$comparison !== 1;
+//          ^ keyword.operator.comparison
+//           ^ keyword.operator.comparison
+//            ^ keyword.operator.comparison
+
+$comparison == 1;
+//          ^ keyword.operator.comparison
+//           ^ keyword.operator.comparison
+
+$comparison != 1;
+//          ^ keyword.operator.comparison
+//           ^ keyword.operator.comparison
+
+$comparison <> 1;
+//          ^ keyword.operator.comparison
+//           ^ keyword.operator.comparison
+
+$comparison <=> 1;
+//          ^ keyword.operator.comparison
+//           ^ keyword.operator.comparison
+//            ^ keyword.operator.comparison