+++ tests/syntax/gt_or_eq_3092/syntax_test_issue_63_fix_assignment_operator.php
@@ -1,0 +1,14 @@
+// SYNTAX TEST "Packages/php-grammar/PHP.tmLanguage"
+<?php
+
+$assignment = 1;
+//          ^ keyword.operator.assignment
+
+$assignment -= 1;
+//          ^ keyword.operator.assignment
+//           ^ keyword.operator.assignment
+
+$assignment <<= 1;
+//          ^ keyword.operator.assignment
+//           ^ keyword.operator.assignment
+//            ^ keyword.operator.assignment