+++ CHANGELOG.md
@@ -8,8 +8,8 @@ I know that many can't update to the latest build. The package will continue wor
 
 * Added: #57 punctuation scope for braces
 * Added: #59 extended echo and print scope: support.function.construct.output
-* Added: extended boolean scope: constant.language.boolean
-* Added: extended magic constant scope: constant.language.magic
+* Added: extended boolean scope to constant.language.boolean
+* Added: extended magic constant scope to constant.language.magic
 * Added: extended null constant scope to "constant.language.null"
 * Fixed: #56 indentation for empty brackets
 * Fixed: #58 syntax bracket handling for closures