+++ CHANGELOG.md
@@ -15,6 +15,7 @@ I know that many can't update to the latest build. The package will continue wor
 * Fixed: #56 indentation for empty brackets
 * Fixed: #58 syntax bracket handling for closures
 * Fixed: #61, #62, and #63 operator scopes not matching correctly
+* Fixed: syntax test suite broken by build 3096
 
 ## 0.16.0